% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/static_paths.R
\name{as.staticPath}
\alias{as.staticPath}
\title{Convert an object to a staticPath object}
\usage{
as.staticPath(path)
}
\arguments{
\item{path}{The object to convert.}
}
\description{
This function is used to convert an object to a \code{staticPath} object.
It is primarily used internally, but can also be used by users to ensure
that an object is a \code{staticPath} object.
}
\keyword{internal}
