% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httpuv.R
\name{startDaemonizedServer}
\alias{startDaemonizedServer}
\title{Create an HTTP/WebSocket daemonized server (deprecated)}
\usage{
startDaemonizedServer(host, port, app, quiet = FALSE)
}
\arguments{
\item{host}{A string that is a valid IPv4 address that is owned by this
server, or \code{"0.0.0.0"} to listen on all IP addresses.}

\item{port}{A number or integer that indicates the server port that should be
listened on. Note that on most Unix-like systems including Linux and macOS,
port numbers smaller than 1024 require root privileges.}

\item{app}{A collection of functions that define your application. See
Details.}

\item{quiet}{If \code{TRUE}, suppress error messages from starting app.}
}
\description{
This function will be removed in a future release of httpuv. It is simply a
wrapper for \code{\link{startServer}}. In previous versions of httpuv (1.3.5
and below), \code{startServer} ran applications in the foreground and
\code{startDaemonizedServer} ran applications in the background, but now both
of them run applications in the background.
}
