% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{stopAllServers}
\alias{stopAllServers}
\title{Stop all servers}
\usage{
stopAllServers()
}
\description{
This will stop all applications which were created by
\code{\link[=startServer]{startServer()}} or \code{\link[=startPipeServer]{startPipeServer()}}.
}
\seealso{
\code{\link[=stopServer]{stopServer()}} to stop a specific server.
}
