% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{encodeURI}
\alias{encodeURI}
\alias{encodeURIComponent}
\alias{decodeURI}
\alias{decodeURIComponent}
\title{URI encoding/decoding}
\usage{
encodeURI(value)

encodeURIComponent(value)

decodeURI(value)

decodeURIComponent(value)
}
\description{
URI encoding/decoding
}
