% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getRNGState}
\alias{getRNGState}
\title{Apply the value of .Random.seed to R's internal RNG state}
\usage{
getRNGState()
}
\description{
This function is needed in unusual cases where a C++ function calls
an R function which sets the value of \code{.Random.seed}. This function
should be called at the end of the R function to ensure that the new value
\code{.Random.seed} is preserved. Otherwise, Rcpp may overwrite it with a
previous value.
}
\keyword{internal}
