% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httpuv-package.R
\docType{package}
\name{httpuv-package}
\alias{httpuv-package}
\alias{httpuv}
\title{HTTP and WebSocket server}
\description{
HTTP and WebSocket server
}
\details{
Allows R code to listen for and interact with HTTP and WebSocket clients, so
you can serve web traffic directly out of your R process. Implementation is
based on \href{https://github.com/joyent/libuv}{libuv} and
\href{https://github.com/nodejs/http-parser}{http-parser}.

This is a low-level library that provides little more than network I/O and
implementations of the HTTP and WebSocket protocols. For an easy way to
create web applications, try \href{https://shiny.posit.co}{Shiny} instead.
}
\examples{
\dontrun{
demo("echo", package="httpuv")
}

}
\seealso{
\link{startServer}
}
\author{
Joe Cheng \email{joe@rstudio.com}
}
\keyword{package}
