% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httpuv2-package.R
\docType{package}
\name{httpuv2-package}
\alias{httpuv2}
\alias{httpuv2-package}
\title{httpuv2: HTTP and WebSocket Server Library}
\description{
Provides low-level socket and protocol support for handling HTTP and WebSocket requests directly from within R. It is primarily intended as a building block for other packages, rather than making it particularly easy to create complete web applications using httpuv alone. httpuv is built on top of the libuv and http-parser C libraries, both of which were developed by Joyent, Inc. (See LICENSE file for libuv and http-parser license information.)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rstudio.github.io/httpuv/}
  \item \url{https://github.com/rstudio/httpuv}
  \item Report bugs at \url{https://github.com/rstudio/httpuv/issues}
}

}
\author{
\strong{Maintainer}: Mauricio Vargas Sepulveda \email{m.vargas.sepulveda@gmail.com} (\href{https://orcid.org/0000-0003-1017-7574}{ORCID})

Authors:
\itemize{
  \item Mauricio Vargas Sepulveda \email{m.vargas.sepulveda@gmail.com} (\href{https://orcid.org/0000-0003-1017-7574}{ORCID})
  \item Joe Cheng \email{joe@posit.co}
  \item Winston Chang \email{winston@posit.co}
  \item Steve Reid (SHA-1 implementation)
  \item James Brown (SHA-1 implementation)
  \item Bob Trower (base64 implementation)
  \item Alexander Peslyak (MD5 implementation)
}

Other contributors:
\itemize{
  \item Posit, PBC (\href{https://ror.org/03wc8by49}{ROR}) [copyright holder, funder]
  \item Hector Corrada Bravo [contributor]
  \item Jeroen Ooms [contributor]
  \item Andrzej Krzemienski (optional.hpp) [copyright holder]
  \item libuv project contributors (libuv library, see src/libuv/AUTHORS file) [copyright holder]
  \item Joyent, Inc. and other Node contributors (libuv library, see src/libuv/AUTHORS file; and http-parser library, see src/http-parser/AUTHORS file) [copyright holder]
  \item Niels Provos (libuv subcomponent: tree.h) [copyright holder]
  \item Internet Systems Consortium, Inc. (libuv subcomponent: inet_pton and inet_ntop, contained in src/libuv/src/inet.c) [copyright holder]
  \item Alexander Chemeris (libuv subcomponent: stdint-msvc2008.h (from msinttypes)) [copyright holder]
  \item Google, Inc. (libuv subcomponent: pthread-fixes.c) [copyright holder]
  \item Sony Mobile Communcations AB (libuv subcomponent: pthread-fixes.c) [copyright holder]
  \item Berkeley Software Design Inc. (libuv subcomponent: android-ifaddrs.h, android-ifaddrs.c) [copyright holder]
  \item Kenneth MacKay (libuv subcomponent: android-ifaddrs.h, android-ifaddrs.c) [copyright holder]
  \item Emergya (Cloud4all, FP7/2007-2013, grant agreement no 289016) (libuv subcomponent: android-ifaddrs.h, android-ifaddrs.c) [copyright holder]
  \item Trantor Standard Systems (base64 implementation) [copyright holder]
  \item Igor Sysoev (http-parser) [copyright holder]
}

}
\keyword{internal}
