% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staticServer.R
\name{runStaticServer}
\alias{runStaticServer}
\title{Serve a directory}
\usage{
runStaticServer(
  dir = getwd(),
  host = "127.0.0.1",
  port = NULL,
  background = FALSE,
  browse = interactive()
)
}
\arguments{
\item{dir}{The directory to serve. Defaults to the current working directory.}

\item{host}{A string that is a valid IPv4 address that is owned by this
server, or \code{"0.0.0.0"} to listen on all IP addresses.}

\item{port}{A number or integer that indicates the server port that should be
listened on. Note that on most Unix-like systems including Linux and Mac OS
X, port numbers smaller than 1025 require root privileges.}

\item{background}{Whether to run the server in the background. By default,
the server runs in the foreground and blocks the R console. You can stop
the server by interrupting it with \code{Ctrl + C}.

When \code{background = TRUE}, the server will run in the background and will
process requests when the R console is idle. To stop a background server,
call \code{\link[=stopAllServers]{stopAllServers()}} or call \code{\link[=stopServer]{stopServer()}} on the server object
returned (invisibly) by this function.}

\item{browse}{Whether to automatically open the served directory in a web
browser. Defaults to \code{TRUE} when running interactively.}
}
\value{
Starts a server on the specified host and port. By default the
server runs in the foreground and is accessible at \verb{http://127.0.0.1:7446}.
When \code{background = TRUE}, the \code{server} object is returned invisibly.
}
\description{
Run a server to host a static directory.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
website_dir <- system.file("example-static-site", package = "httpuv")
runStaticServer(dir = website_dir)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=runServer]{runServer()}} provides a similar interface for running a dynamic
app server. Both \code{runStaticServer()} and \code{\link[=runServer]{runServer()}} are built on top of
\code{\link[=startServer]{startServer()}}, \code{\link[=service]{service()}} and \code{\link[=stopServer]{stopServer()}}. Learn more about httpuv
servers in \code{\link[=startServer]{startServer()}}.
}
