% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/static_paths.R
\name{staticPath}
\alias{staticPath}
\title{Create a staticPath object}
\usage{
staticPath(path, index = TRUE, fallthrough = FALSE)
}
\arguments{
\item{path}{The local path.}

\item{index}{If an index.html file is present, should it be served up when
the client requests \code{/} or any subdirectory?}

\item{fallthrough}{With the default value, \code{FALSE}, if a request is made
for a file that doesn't exist, then httpuv will immediately send a 404
response from the background I/O thread, without needing to call back into
the main R thread. This offers the best performance. If the value is
\code{TRUE}, then instead of sending a 404 response, httpuv will call the
application's \code{call} function, and allow it to handle the request.}
}
\description{
This function creates a \code{staticPath} object.
}
