% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httpuv.R
\name{stopDaemonizedServer}
\alias{stopDaemonizedServer}
\title{Stop a running daemonized server in Unix environments (deprecated)}
\usage{
stopDaemonizedServer(handle)
}
\arguments{
\item{handle}{A handle that was previously returned from
\code{\link{startServer}} or \code{\link{startPipeServer}}.}
}
\description{
This function will be removed in a future release of httpuv. Instead, use
\code{\link{stopServer}}.
}
