% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{stopServer}
\alias{stopServer}
\title{Stop a server}
\usage{
stopServer(server)
}
\arguments{
\item{server}{A server object that was previously returned from
\code{\link[=startServer]{startServer()}} or \code{\link[=startPipeServer]{startPipeServer()}}.}
}
\description{
Given a server object that was returned from a previous invocation of
\code{\link[=startServer]{startServer()}} or \code{\link[=startPipeServer]{startPipeServer()}}, this closes all
open connections for that server and unbinds the port.
}
\seealso{
\code{\link[=stopAllServers]{stopAllServers()}} to stop all servers.
}
