% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{stopServer}
\alias{stopServer}
\title{Stop a server}
\usage{
stopServer(server)
}
\arguments{
\item{handle}{A handle that was previously returned from
\code{\link{startServer}} or \code{\link{startPipeServer}}.}
}
\description{
Given a handle that was returned from a previous invocation of 
\code{\link{startServer}} or \code{\link{startPipeServer}}, this closes all
open connections for that server and  unbinds the port.
}
\seealso{
\code{\link{stopAllServers}} to stop all servers.
}
